removed unused utilities
- the tests where moved to the raiden-contracts repo, so the utilities to
find the smart contract are not required
- the tests don't need to query the blockchain anymore to get the
events, since they are available in the WAL, so the
get_channel_events_for_token is not necessary anymore
- dump_messages is not necessary anymore because of the WAL
